+++ sdk/metricsadvisor/ai-metrics-advisor/recordings/node/aad_metricsadvisorclient/recording_listmetricseriesdefinitions_by_page.js
@@ -1,0 +1,82 @@
+let nock = require('nock');
+
+module.exports.hash = "c2f50ecc051b8d62b645034e2bdced2f";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}
+
+nock('https://login.microsoftonline.com:443', {"encodedQueryParams":true})
+  .post('/azure_tenant_id/oauth2/v2.0/token', "response_type=token&grant_type=client_credentials&client_id=azure_client_id&client_secret=azure_client_secret&scope=https%3A%2F%2Fcognitiveservices.azure.com%2F.default")
+  .reply(200, {"token_type":"Bearer","expires_in":86399,"ext_expires_in":86399,"access_token":"access_token"}, [
+  'Cache-Control',
+  'no-store, no-cache',
+  'Pragma',
+  'no-cache',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Expires',
+  '-1',
+  'Strict-Transport-Security',
+  'max-age=31536000; includeSubDomains',
+  'X-Content-Type-Options',
+  'nosniff',
+  'P3P',
+  'CP="DSP CUR OTPi IND OTRi ONL FIN"',
+  'x-ms-request-id',
+  '7ffc9287-9082-4afd-a322-06dbb3454800',
+  'x-ms-ests-server',
+  '2.1.11397.13 - SEASLR1 ProdSlices',
+  'Set-Cookie',
+  'fpc=AsoDeAShyiJLlORVED-UD3HGLH8mEAAAAGdLk9cOAAAA; expires=Sun, 14-Feb-2021 08:40:29 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'Set-Cookie',
+  'x-ms-gateway-slice=estsfd; path=/; secure; samesite=none; httponly',
+  'Set-Cookie',
+  'stsservicecookie=estsfd; path=/; secure; samesite=none; httponly',
+  'Date',
+  'Fri, 15 Jan 2021 08:40:29 GMT',
+  'Content-Length',
+  '1331'
+]);
+
+nock('https://endpoint:443', {"encodedQueryParams":true})
+  .post('/metricsadvisor/v1.0/metrics/189ff959-d9f4-45c7-a1e0-f87c9c7ca80f/series/query', {"activeSince":"2020-08-05T00:00:00.000Z"})
+  .query(true)
+  .reply(200, {"value":[{"metricId":"189ff959-d9f4-45c7-a1e0-f87c9c7ca80f","dimension":{"city":"Miami","category":"Health & Personal Care"}},{"metricId":"189ff959-d9f4-45c7-a1e0-f87c9c7ca80f","dimension":{"city":"Rio de Janeiro","category":"Historical & Advertising Collectibles"}}],"@nextLink":"https://endpoint:443/metricsadvisor/v1.0/metrics/189ff959-d9f4-45c7-a1e0-f87c9c7ca80f/series/query?$top=2&$skip=2"}, [
+  'Content-Length',
+  '435',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'x-request-id',
+  '70abe16c-29ef-47b5-b2f9-bb912471855b',
+  'x-envoy-upstream-service-time',
+  '61',
+  'apim-request-id',
+  '70abe16c-29ef-47b5-b2f9-bb912471855b',
+  'Strict-Transport-Security',
+  'max-age=31536000; includeSubDomains; preload',
+  'x-content-type-options',
+  'nosniff',
+  'Date',
+  'Fri, 15 Jan 2021 08:40:29 GMT'
+]);
+
+nock('https://endpoint:443', {"encodedQueryParams":true})
+  .post('/metricsadvisor/v1.0/metrics/189ff959-d9f4-45c7-a1e0-f87c9c7ca80f/series/query', {"activeSince":"2020-08-05T00:00:00.000Z"})
+  .query(true)
+  .reply(200, {"value":[{"metricId":"189ff959-d9f4-45c7-a1e0-f87c9c7ca80f","dimension":{"city":"Belo Horizonte","category":"Electronics (Consumer)"}},{"metricId":"189ff959-d9f4-45c7-a1e0-f87c9c7ca80f","dimension":{"city":"Kinshasa","category":"Beauty"}}],"@nextLink":"https://endpoint:443/metricsadvisor/v1.0/metrics/189ff959-d9f4-45c7-a1e0-f87c9c7ca80f/series/query?$top=2&$skip=4"}, [
+  'Content-Length',
+  '407',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'x-request-id',
+  'a50238a7-ffb7-44f6-8ea7-fc3c6a44302d',
+  'x-envoy-upstream-service-time',
+  '58',
+  'apim-request-id',
+  'a50238a7-ffb7-44f6-8ea7-fc3c6a44302d',
+  'Strict-Transport-Security',
+  'max-age=31536000; includeSubDomains; preload',
+  'x-content-type-options',
+  'nosniff',
+  'Date',
+  'Fri, 15 Jan 2021 08:40:29 GMT'
+]);