+++ sdk/eventhub/event-hubs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 5.4.0 (Unreleased)
 
+- A helper method `parseEventHubConnectionString` has been added which validates and
+  parses a given connection string for Azure Event Hubs.
+  Resolves [#11894](https://github.com/Azure/azure-sdk-for-js/issues/11894)
+
 - Adds the `customEndpointAddress` field to `EventHubClientOptions`.
   This allows for specifying a custom endpoint to use when communicating
   with the Event Hubs service, which is useful when your network does not